refactor: migrate tile event handlers to `Server.Events` and update namespaces
- Moved `TileChangedEventHandler` and `TileLoadedEventHandler` from `World.Map` to `Server.Events.World`.
- Updated related namespaces and project structure to align with the new module.
- Adjusted test files and dependencies to reflect the migration.
