+++ tests/NeoServer.Domain.Tests/Creature/Players/PlayerUseTests.cs
@@ -8,6 +8,7 @@ using NeoServer.Domain.World.Events;
 using NeoServer.Domain.World.Map;
 using NeoServer.Domain.World.Models.Tiles;
 using NeoServer.Server.Commands.Movements;
+using NeoServer.Server.Events.World;
 
 namespace NeoServer.Domain.Tests.Creature.Players;
